Create debug files on --debug
