+++ src/zap_hooks/helpers/configuration.py
@@ -36,6 +36,7 @@ class DASTConfig:
     oauth_token_url: Optional[str] = None
     oauth_parameters: Optional[str] = None
     disable_rules: Optional[str] = None
+    debug_mode: Optional[bool] = False
 
     def __init__(self):
         self.extra_zap_params = None
@@ -71,6 +72,7 @@ class DASTConfig:
             self.oauth_token_url = os.environ.get('OAUTH_TOKEN_URL') or EMPTY_STRING
             self.oauth_parameters = self._get_hook_param_list(os.environ.get('OAUTH_PARAMETERS')) or EMPTY_STRING
             self.disable_rules = self._get_hook_param_list(os.environ.get('DISABLE_RULES')) or None
+            self.debug_mode = os.environ.get('DEBUG_MODE') or False
 
         except Exception as error:
             log(f"error in start_docker_zap: {traceback.print_exc()}", log_level=LogLevel.ERROR)
