+++ src/zap_hooks/helpers/utilities.py
@@ -1,3 +1,4 @@
+import json
 from sys import exit
 from typing import Dict, Iterable, NoReturn
 
@@ -37,3 +38,26 @@ def array_to_dict(array: Iterable[str]):
         key, value = key_value.split(':', 1)
         body.append((key, value))
     return body
+
+def serialize_and_save(obj, filename):
+    serialized_data = serialize_object(obj)
+    with open(filename, 'w') as file:
+        json.dump(serialized_data, file, indent=4)
+
+def serialize_object(obj):
+    serialized_data = {}
+    for attr in dir(obj):
+        value = getattr(obj, attr)
+        if is_serializable(value):
+            serialized_data[attr] = value
+        else:
+            pass
+    return serialized_data
+
+def is_serializable(value):
+    try:
+        json.dumps(value)
+        return True
+    except (TypeError, OverflowError):
+        return False
+
